+++ app/src/main/java/mb28/crysongs/ui/MiniPlayer.kt
@@ -13,33 +13,25 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxHeight
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.CircularWavyProgressIndicator
-import androidx.compose.material3.FloatingActionButton
-import androidx.compose.material3.FloatingActionButtonDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
+import androidx.compose.material3.ToggleButton
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.text.style.LineHeightStyle
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
@@ -51,7 +43,6 @@ import mb28.crysongs.core.formatDurationMs
 import mb28.crysongs.duration
 import mb28.crysongs.icons.pause_circle
 import mb28.crysongs.icons.play_circle
-import mb28.crysongs.icons.search
 import mb28.crysongs.icons.skip_next
 import mb28.crysongs.icons.skip_previous
 import mb28.crysongs.icons.swap_horiz
@@ -72,10 +63,9 @@ fun MiniPlayer(secondSet: MutableState<Boolean>) {
         verticalAlignment = Alignment.CenterVertically,
         horizontalArrangement = Arrangement.SpaceBetween
     ) {
-        FloatingActionButton(
-            { secondSet.value = !secondSet.value },
-            elevation = FloatingActionButtonDefaults.elevation(
-                0.dp, 0.dp)
+        ToggleButton(
+            secondSet.value,
+            { secondSet.value = !secondSet.value }
         ) {
             Icon(swap_horiz, null)
         }